move det_terminal_size to systemtools.py
